moved grouping module into pauli
